// apps/gq/src/app/calculator/rfq-4-overview-view/components/rfq-4-requests-table/configs/rfq-4-requests-fields.enum.ts
export enum Rfq4RequestsFields {
  GQ_POSITION_ID = 'gqPositionId',
  RFQ_ID = 'rfqId',
  MAT_DESC = 'materialDesc',
  MAT_NUMBER = 'materialNumber',
  STATUS = 'calculatorRequestRecalculationStatus',
  CUST_NAME = 'customerName',
  CUST_NUMBER = 'customerNumber',
  CREATED_BY = 'createdBy',
  LAST_UPDATED = 'rfqLastUpdated',
  ASSIGNED_TO = 'assignedTo',
}